VFS and SCAFS updates for OSSIE integration. "list" function is now generic, not per-fs impl. adapt/os/ocpios/interfaces/include/OcpiOsFileSystem.h: add Dir object for reading directories easier adapt/os/ocpios/linux/src/OcpiOsFileSystem.cxx: ditto, use namespace decls in source file core/local/util/bin/src/HttpFsDemo_main.cxx: update to use new generic vfs lister, supplying dir reader core/local/util/bin/src/HttpServerDemo_main.cxx: update to use default FileFs root (cwd) core/local/util/bin/src/ZipReaderDemo_main.cxx: update to use new generic vfs lister core/local/util/bin/src/ZipWriterDemo_main.cxx: ditto core/local/util/filefs/include/OcpiUtilFileFs.h: allow default constructor on cwd, new dir reader core/local/util/filefs/src/OcpiUtilFileFs.cxx: ditto, and use namespace decls core/local/util/http/include/OcpiUtilHttpFs.h: new dir reader core/local/util/http/src/OcpiUtilHttpFs.cxx: ditto core/local/util/memfs/src/OcpiUtilStaticMemFs.cxx: remove old iterator core/local/util/streamfs/src/OcpiUtilStreamFsReader.cxx: remove old iterator core/local/util/vfs/include/OcpiUtilUriFs.h: ditto, and add new dir reader core/local/util/vfs/include/OcpiUtilVfs.h: new lister, virtual dir reader interface core/local/util/vfs/include/OcpiUtilVfsIterator.h: new lister core/local/util/vfs/src/OcpiUtilFilterFs.cxx: remove old iterator core/local/util/vfs/src/OcpiUtilUriFs.cxx: new dir reader, remove old iterator core/local/util/vfs/src/OcpiUtilVfs.cxx: use namespace decls, new lister, push/pop recursive dir reader core/local/util/vfs/src/OcpiUtilVfsIterator.cxx: new generic lister core/local/util/vfs/src/OcpiUtilVfsUtil.cxx: use new lister core/local/util/zipfs/include/OcpiUtilZipFs.h: new dir reader, remove old iterator core/local/util/zipfs/src/OcpiUtilZipFs.cxx: ditto core/sca/cf_util/bin/src/loadWorker_main.cxx: remove unneeded orb arg to SCAfs constructor core/sca/cf_util/vfs/include/OcpiCFUtilSCAFs.h: new dir reader, remove old iterator core/sca/cf_util/vfs/src/OcpiCFUtilSCAFs.cxx: ditto core/sca/cf_util/vfs/include/OcpiCFUtilVfsFileSystem.h: remove orb arg for constructor, virtually inherit POA base core/sca/cf_util/vfs/src/OcpiCFUtilVfsFileSystem.cxx: ditto, and user new iterator

+++ core/local/util/vfs/src/OcpiUtilVfsUtil.cxx
@@ -32,12 +32,13 @@
  *  along with OpenCPI.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <OcpiUtilVfsUtil.h>
-#include <OcpiUtilVfs.h>
-#include <OcpiUtilVfsIterator.h>
-#include <OcpiUtilMisc.h>
 #include <istream>
 #include <string>
+
+#include "OcpiUtilVfsUtil.h"
+#include "OcpiUtilVfs.h"
+#include "OcpiUtilVfsIterator.h"
+#include "OcpiUtilMisc.h"
 
 /*
  * ----------------------------------------------------------------------
@@ -95,59 +96,22 @@ copyFilesRecursively (Vfs * srcFs, const std::string & srcDir,
                       const std::string & pattern)
   throw (std::string)
 {
-  if (OCPI::Util::Vfs::relativeName (pattern) != pattern) {
+  if (OCPI::Util::Vfs::relativeName (pattern) != pattern)
     throw std::string ("pattern must not contain directory separator");
+
+  Iterator it(*srcFs, srcDir, pattern.c_str());
+  bool isDir;
+  std::string name;
+
+  while (it.next(name, isDir)) {
+    std::string
+      srcName(joinNames(srcDir, name)),
+      destName(joinNames(destDir, name));
+    if (isDir)
+      copyDirectoryRecursively(srcFs, srcName,	destFs, destName, pattern);
+    else
+      srcFs->copy(srcName, destFs, destName);
   }
-
-  /*
-   * Iterate over all subdirectories
-   */
-
-  OCPI::Util::Vfs::Iterator * it = srcFs->list (srcDir);
-
-  try {
-    while (!it->end()) {
-      if (it->isDirectory()) {
-        std::string newSrcDir =
-          OCPI::Util::Vfs::joinNames (srcDir, it->relativeName());
-        std::string newDestDir =
-          OCPI::Util::Vfs::joinNames (destDir, it->relativeName());
-        copyDirectoryRecursively (srcFs, newSrcDir,
-                                  destFs, newDestDir,
-                                  pattern);
-      }
-      it->next ();
-    }
-  }
-  catch (...) {
-    srcFs->closeIterator (it);
-    throw;
-  }
-
-  srcFs->closeIterator (it);
-
-  /*
-   * Copy all matching files
-   */
-
-  it = srcFs->list (srcDir, pattern);
-
-  try {
-    while (!it->end()) {
-      if (!it->isDirectory()) {
-        std::string destFileName =
-          OCPI::Util::Vfs::joinNames (destDir, it->relativeName());
-        srcFs->copy (it->absoluteName(), destFs, destFileName);
-      }
-      it->next ();
-    }
-  }
-  catch (...) {
-    srcFs->closeIterator (it);
-    throw;
-  }
-
-  srcFs->closeIterator (it);
 }
 
 void
@@ -169,73 +133,33 @@ removeFilesRecursively (Vfs * fs,
                         bool deleteEmptyDirectories)
   throw (std::string)
 {
-  if (OCPI::Util::Vfs::relativeName (pattern) != pattern) {
+  if (OCPI::Util::Vfs::relativeName (pattern) != pattern)
     throw std::string ("pattern must not contain directory separator");
-  }
 
-  /*
-   * Iterate over all subdirectories
-   */
+  Iterator it(*fs, dir, pattern.c_str());
+  bool isDir;
+  std::string name;
 
-  OCPI::Util::Vfs::Iterator * it = fs->list (dir);
-
-  try {
-    while (!it->end()) {
-      if (it->isDirectory()) {
-        removeFilesRecursively (fs, it->absoluteName(), pattern,
-                                deleteEmptyDirectories);
-
-        /*
-         * If deleteEmptyDirectories is true, see if that directory
-         * is now empty. If yes, delete it.
-         */
-
-        if (deleteEmptyDirectories) {
-          OCPI::Util::Vfs::Iterator * it2 =
-            fs->list (it->absoluteName());
-
-          try {
-            it2->end ();
-          }
-          catch (...) {
-            fs->closeIterator (it2);
-            throw;
-          }
-
-          fs->closeIterator (it2);
-          fs->rmdir (it->absoluteName());
-        }
+  while (it.next(name, isDir)) {
+    std::string fullName = joinNames(dir, name);
+    if (isDir) {
+      removeFilesRecursively(fs, fullName, pattern, deleteEmptyDirectories);
+      /*
+       * If deleteEmptyDirectories is true, see if that directory
+       * is now empty. If yes, delete it.
+       */
+      if (deleteEmptyDirectories) {
+	bool empty;
+	{
+	  Iterator it2(*fs, fullName);
+	  empty = !it2.next(name, isDir);
+	}
+	if (empty)
+          fs->rmdir(fullName);
       }
-      it->next ();
-    }
+    } else
+      fs->remove(fullName);
   }
-  catch (...) {
-    fs->closeIterator (it);
-    throw;
-  }
-
-  fs->closeIterator (it);
-
-  /*
-   * Delete all matching files
-   */
-
-  it = fs->list (dir, pattern);
-
-  try {
-    while (!it->end()) {
-      if (!it->isDirectory()) {
-        fs->remove (it->absoluteName());
-      }
-      it->next ();
-    }
-  }
-  catch (...) {
-    fs->closeIterator (it);
-    throw;
-  }
-
-  fs->closeIterator (it);
 }
 
 void
@@ -249,18 +173,14 @@ removeDirectoryRecursively (Vfs * fs,
   removeFilesRecursively (fs, dir, pattern, deleteEmptyDirectories);
 
   if (deleteEmptyDirectories) {
-    OCPI::Util::Vfs::Iterator * it = fs->list (dir);
-
-    try {
-      it->end ();
+    std::string name;
+    bool isDir, empty;
+    {
+      Iterator it(*fs, dir);
+      empty = !it.next(name, isDir);
     }
-    catch (...) {
-      fs->closeIterator (it);
-      throw;
-    }
-
-    fs->closeIterator (it);
-    fs->rmdir (dir);
+    if (empty)
+      fs->rmdir(dir);
   }
 }
